functions/video-subtitles/requirements.txt
functions-framework==3.*
google-cloud-storage==2.14.0
faster-whisper==1.0.3
